requirements.txt
pytest == 7.1.1
selenium==4.1.3
mysql-connector-python==8.0.22
browsermob-proxy==0.8.0
mitmproxy==7.0.4
requests == 2.27.1
cerberus == 1.3.4
webdriver-manager == 3.5.4
Faker == 13.11.0
allure-pytest == 2.9.45
pytest-xdist == 2.5.0
paramiko == 2.11.0
#pytest-failed-screenshot ~= 1.0.2
